+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.1.15 - Unreleased
 
+### Added
+* Add yes argument to CLI
+
 
 ## Version 0.1.14 - Released 2023-05-11