+++ CHANGELOG.md
@@ -5,7 +5,16 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.1.14 - Unreleased
+## Version 0.1.15 - Unreleased
+
+### Added
+* Add yes argument to CLI
+
+### Changed
+* Added more options to the serial queue `run` method.
+
+
+## Version 0.1.14 - Released 2023-05-11
 
 
 ## Version 0.1.13 - Released 2023-05-11